refactor: AgeDiscount, DistanceDiscount 리팩토링
- 패키지 위치 변경
- DistanceDecorator 네이밍 변경
- AgeDiscount extraFare 계산 로직 추상화
