+++ src/main/java/nextstep/subway/path/domain/age/FareDeductible.java
@@ -1,0 +1,5 @@
+package nextstep.subway.path.domain.age;
+
+public interface FareDeductible {
+    int calculateDeductedFare(int originalFare);
+}
