+++ src/main/java/nextstep/subway/path/domain/distance/DistanceCalculate.java
@@ -1,0 +1,14 @@
+package nextstep.subway.path.domain.distance;
+
+public abstract class DistanceCalculate implements DistancePolicy {
+
+    private final DistancePolicy policy;
+
+    protected DistanceCalculate(DistancePolicy policy) {
+        this.policy = policy;
+    }
+
+    protected int delegate() {
+        return policy.calculate();
+    }
+}
